#233 refactor: 블록 선택시 로직 커스텀 훅으로 분리
