+++ src/page/archiving/index/ArchivingPage.tsx
@@ -1,4 +1,4 @@
-import { useEffect, useMemo, useRef, useState } from 'react';
+import { useEffect, useRef } from 'react';
 import { useLocation } from 'react-router-dom';
 
 import Add from '@/common/asset/svg/ic_add_btn.svg?react';
@@ -19,79 +19,46 @@ import DocumentBar from '@/page/archiving/index/component/DocumentBar/DocumentBa
 import MonthHeader from '@/page/archiving/index/component/MonthHeader/MonthHeader';
 import TimeBlock from '@/page/archiving/index/component/TimeBlock/TimeBlock';
 import YearHeader from '@/page/archiving/index/component/YearHeader/YearHeader';
-import { useGetTimeBlockQuery } from '@/page/archiving/index/hook/api/useGetTimeBlockQuery';
 import { useDate } from '@/page/archiving/index/hook/common/useDate';
+import { useFetchBlock } from '@/page/archiving/index/hook/common/useFetchBlock';
+import { useSelectBlock } from '@/page/archiving/index/hook/common/useSelectBlock';
 import { Block } from '@/page/archiving/index/type/blockType';
 import { MonthType } from '@/page/archiving/index/type/monthType';
-import { alignBlocks, createTimeBlock } from '@/page/archiving/index/util/block';
-
-import { useOpenModal } from '@/shared/store/modal';
+import { createTimeBlock } from '@/page/archiving/index/util/block';
 
 const ArchivingPage = () => {
-  const [selectedBlock, setSelectedBlock] = useState<Block>();
+  const location = useLocation();
 
   const daySectionRef = useRef<HTMLDivElement>(null);
 
-  const location = useLocation();
+  const sideBarRef = useOutsideClick(() => setSelectedBlock(undefined));
 
   const teamId = new URLSearchParams(location.search).get('teamId');
-  if (!teamId) throw new Error('has no error');
+  if (!teamId) throw new Error('has no teamId');
 
   const {
-    currentDate,
     currentYear,
     selectedMonthString,
     setSelectedMonthString,
     handlePrevYear,
     handleNextYear,
     endDay,
+    handleMonthClick,
   } = useDate(daySectionRef);
-  const selectedMonth = parseInt(selectedMonthString.split('월')[0]);
 
-  const { data } = useGetTimeBlockQuery(+teamId, 'executive', currentYear, selectedMonth);
+  const { selectedBlock, setSelectedBlock, handleOpenBlockModal, handleBlockClick } = useSelectBlock();
 
-  // eslint-disable-next-line react-hooks/exhaustive-deps
-  const timeBlocks: Block[] = data?.timeBlocks || [];
-  const blockFloors = useMemo(
-    () => alignBlocks(timeBlocks, endDay, selectedMonthString, currentYear),
-    [currentYear, endDay, selectedMonthString, timeBlocks]
-  );
-
-  const handleClose = () => {
-    selectedBlock && setSelectedBlock(undefined);
-  };
-  const sideBarRef = useOutsideClick(handleClose);
-
-  const openModal = useOpenModal();
+  const { timeBlocks, blockFloors } = useFetchBlock({
+    teamId: +teamId,
+    currentYear,
+    selectedMonth: +selectedMonthString.split('월')[0],
+    endDay,
+    selectedMonthString,
+  });
 
   useEffect(() => {
-    setSelectedMonthString(`${currentDate.getMonth() + 1}월` as MonthType);
-    // eslint-disable-next-line react-hooks/exhaustive-deps
+    setSelectedMonthString(`${new Date().getMonth() + 1}월` as MonthType);
   }, [teamId]);
-
-  const handleOpenBlockModal = () => {
-    openModal('create-block');
-  };
-
-  const handleMonthClick = (month: MonthType) => {
-    daySectionRef.current?.scrollTo(0, 0);
-    setSelectedMonthString(month);
-  };
-
-  const handleBlockClick = (
-    e: React.MouseEvent<HTMLDivElement> | React.KeyboardEvent<HTMLDivElement>,
-    block: Block
-  ) => {
-    e.stopPropagation();
-
-    e.currentTarget.scrollIntoView({
-      behavior: 'smooth',
-      inline: 'center',
-      block: 'center',
-    });
-
-    setSelectedBlock(block);
-  };
 
   return (
     <Flex css={pageStyle}>
@@ -112,7 +79,7 @@ const ArchivingPage = () => {
                 startDate: new Date(startDate),
                 endDate: new Date(endDate),
                 currentYear,
-                selectedMonth: selectedMonth,
+                selectedMonth: +selectedMonthString.split('월')[0],
               });
 
               return (
@@ -139,7 +106,7 @@ const ArchivingPage = () => {
           </Button>
         </Flex>
       </section>
-      <DocumentBar selectedBlock={selectedBlock} ref={sideBarRef} onClose={handleClose} />
+      <DocumentBar selectedBlock={selectedBlock} ref={sideBarRef} onClose={() => setSelectedBlock(undefined)} />
     </Flex>
   );
 };
